Add codebase setup essential files
Added files and configurations that form the essential structure of the current Scala project. This includes configuration files for the Scala formatter, AWS and IntelliJ settings, SBT build properties and plugins, Scala compiler settings, and importantly, the Code of Conduct and License files. This change helps to ensure a consistent development environment and understanding of code responsibilities and behavior expectations when contributing to this project.