Add GameCharacter class and corresponding test
Introduced a new 'GameCharacter' class in the Scala project, which has properties like 'name' and 'hp'. Also, introduced a corresponding test suite 'GameCharacterTest' that asserts the creation of a new instance of 'GameCharacter'. With these changes, it would be easier to manage and manipulate game character related data throughout the game's lifecycle.