Refactor game characters to traits and abstract classes
Refactored the GameCharacter class into a trait and created an AbstractCharacter abstract class. Now, the game characters such as Mage and Warrior, are extending from AbstractCharacter which itself extends from GameCharacter trait. The logic of keeping current hit points is moved to this AbstractCharacter class. This increases code reuse and allows better control and flexibility of character behavior and attributes. Also, a method for attacking is defined in the GameCharacter trait to force all characters to implement their version of attacks.